Refactor: RootLayout 조건부 스타일 정리 및 아이콘 네이밍 수정
